Prevent pactman-verifier from crashing for unhandled pact states
When running pactman-verifier, if the request to set up provider state
returned an error pactman would attempt to log "Invalid provider state".
However because the CaptureResult class did not implement the warn
method, the Result superclass would raise a NotImplementedError. This
would crash pactman-verifier. This change makes CaptureResult log the
warning message.
