+++ .gitignore
@@ -91,3 +91,5 @@ ENV/
 
 # Rope project settings
 .ropeproject
+
+/*.json
