+++ pactman/__version__.py
@@ -1,1 +1,1 @@
-__version__ = '2.20.0'
+__version__ = '2.20.1'
