+++ pactman/verifier/broker_pact.py
@@ -3,7 +3,6 @@
 https://github.com/pact-foundation/pact-specification/tree/version-2
 """
 import json
-import logging
 import os
 import urllib.parse
 
@@ -12,9 +11,6 @@ from restnavigator import Navigator
 
 from .result import LoggedResult
 from .verify import Interaction
-
-log = logging.getLogger(__name__)
-log.setLevel(logging.DEBUG)
 
 
 def pact_id(param):
