+++ pactman/verifier/result.py
@@ -1,5 +1,7 @@
 import logging
-from colorama import Fore
+from colorama import Fore, Style
+
+from .paths import format_path
 
 log = logging.getLogger(__name__)
 
@@ -49,3 +51,47 @@ class PytestResult(Result):   # pragma: no cover
         self.success = self.FAIL
         log.warning(' ' + message)
         raise Failed(message) from None
+
+
+class CaptureResult(Result):
+    def __init__(self, *, level=logging.WARNING):
+        self.messages = []
+        self.level = level
+        self.current_consumer = None
+
+    def start(self, interaction):
+        super().start(interaction)
+        log = logging.getLogger('pactman')
+        log.handlers = [self]
+        log.setLevel(logging.DEBUG)
+        self.messages[:] = []
+        if self.current_consumer != interaction.pact.consumer:
+            print(f'{Style.BRIGHT}Consumer: {interaction.pact.consumer}')
+            self.current_consumer = interaction.pact.consumer
+        print(f'Request: "{interaction.description}" ... ', end='')
+
+    def end(self):
+        if self.success:
+            print(Fore.GREEN + 'PASSED')
+        else:
+            print(Fore.RED + 'FAILED')
+        if self.messages:
+            print((Fore.RESET + '\n').join(self.messages))
+
+    def warn(self, message):
+        log.warning(message)
+
+    def fail(self, message, path=None):
+        self.success = self.FAIL
+        if path:
+            message += ' at ' + format_path(path)
+        log.error(message)
+        return not message
+
+    def handle(self, record):
+        color = ''
+        if record.levelno >= logging.ERROR:
+            color = Fore.RED
+        elif record.levelno >= logging.WARNING:
+            color = Fore.YELLOW
+        self.messages.append(' ' + color + record.msg)
